Fix terrible glue code that doesn't work for identity rotation.
This glue code converted from one quaternion class to another via
conversion to and from axis-angle representations.  Instead, it could have
just used getters/setters to transfer the w,x,y,z components (as the fix
does).

The axis-angle approach failed in the identity transformation case -- it
resulted in NAN somewhere that caused problems down the chain.  This
implies that one quaternion approach or another has a broken axis-angle
getter / setter.  It would be best to fix that tooquaternion.